Added move validation

+++ raptor/src/raptor/game/util/ZobristHash.java
@@ -6,21 +6,22 @@ import java.util.Random;
 import raptor.game.Game;
 import raptor.game.GameConstants;
 
+
 import static raptor.game.util.GameUtils.*;
 
 public final class ZobristHash implements GameConstants {
 
-	private static int[][][] ZOBRIST_POSITION = new int[2][7][64];
-	private static int[] ZOBRIST_TO_MOVE = new int[2];
-	private static int[] ZOBRIST_EP = new int[65];
-	private static int[][] ZOBRIST_CASTLE = new int[2][4];
-	private static int[][][] ZOBRIST_DROP_COUNT = new int[2][7][18];
+	private static long[][][] ZOBRIST_POSITION = new long[2][7][64];
+	private static long[] ZOBRIST_TO_MOVE = new long[2];
+	private static long[] ZOBRIST_EP = new long[65];
+	private static long[][] ZOBRIST_CASTLE = new long[2][4];
+	private static long[][][] ZOBRIST_DROP_COUNT = new long[2][7][18];
 
 	static {
 		initZobrist();
 	}
 
-	public static int zobristHashPositionOnly(Game game) {
+	public static long zobristHashPositionOnly(Game game) {
 		return zobristPiece(WHITE, PAWN, game)
 				^ zobristPiece(WHITE, BISHOP, game)
 				^ zobristPiece(WHITE, KNIGHT, game)
@@ -35,7 +36,7 @@ public final class ZobristHash implements GameConstants {
 				^ zobristPiece(BLACK, KING, game);
 	}
 
-	private static int zobristPiece(int color, int piece, Game game) {
+	private static long zobristPiece(int color, int piece, Game game) {
 		int result = 0;
 		long current = game.getPieceBB(color, piece);
 		while (current != 0L) {
@@ -45,24 +46,24 @@ public final class ZobristHash implements GameConstants {
 		return result;
 	}
 
-	public static int zobristHash(Game game) {
+	public static long zobristHash(Game game) {
 		return zobristHashPositionOnly(game)
 				^ zobrist(game.getColorToMove(), game.getEpSquare(), game
 						.getCastling(WHITE), game.getCastling(BLACK));
 	}
 
-	public static int zobrist(int color, int piece, int square) {
+	public static long zobrist(int color, int piece, int square) {
 		return ZOBRIST_POSITION[color][piece][square];
 	}
 
-	public static int zobrist(int colorToMove, int epSquare, int whiteCastling,
+	public static long zobrist(int colorToMove, int epSquare, int whiteCastling,
 			int blackCastling) {
 		return ZOBRIST_TO_MOVE[colorToMove] ^ ZOBRIST_EP[epSquare]
 				^ ZOBRIST_CASTLE[WHITE][whiteCastling]
 				^ ZOBRIST_CASTLE[BLACK][blackCastling];
 	}
 
-	public static int zobristDropPieces(Game game) {
+	public static long zobristDropPieces(Game game) {
 		return ZOBRIST_DROP_COUNT[WHITE][PAWN][game.getDropCount(WHITE, PAWN)]
 				^ ZOBRIST_DROP_COUNT[WHITE][PAWN][game.getDropCount(WHITE,
 						KNIGHT)]
@@ -90,7 +91,7 @@ public final class ZobristHash implements GameConstants {
 		for (int i = 0; i < ZOBRIST_DROP_COUNT.length; i++) {
 			for (int j = 0; j < ZOBRIST_DROP_COUNT[i].length; j++) {
 				for (int k = 0; k < ZOBRIST_DROP_COUNT[i][j].length; k++) {
-					ZOBRIST_DROP_COUNT[i][j][k] = random.nextInt();
+					ZOBRIST_DROP_COUNT[i][j][k] = random.nextLong();
 				}
 			}
 		}
@@ -98,22 +99,22 @@ public final class ZobristHash implements GameConstants {
 		for (int i = 0; i < ZOBRIST_POSITION.length; i++) {
 			for (int j = 0; j < ZOBRIST_POSITION[i].length; j++) {
 				for (int k = 0; k < ZOBRIST_POSITION[i][j].length; k++) {
-					ZOBRIST_POSITION[i][j][k] = random.nextInt();
+					ZOBRIST_POSITION[i][j][k] = random.nextLong();
 				}
 			}
 		}
 
 		for (int i = 0; i < ZOBRIST_TO_MOVE.length; i++) {
-			ZOBRIST_TO_MOVE[i] = random.nextInt();
+			ZOBRIST_TO_MOVE[i] = random.nextLong();
 		}
 
 		for (int i = 0; i < ZOBRIST_EP.length; i++) {
-			ZOBRIST_EP[i] = random.nextInt();
+			ZOBRIST_EP[i] = random.nextLong();
 		}
 
 		for (int i = 0; i < ZOBRIST_CASTLE.length; i++) {
 			for (int j = 0; j < ZOBRIST_CASTLE[i].length; j++) {
-				ZOBRIST_CASTLE[i][j] = random.nextInt();
+				ZOBRIST_CASTLE[i][j] = random.nextLong();
 			}
 		}
 	}
